chore: move square files
