+++ src/model/game.ts
@@ -4,7 +4,7 @@ import { Deck } from "./deck";
 import { Dice } from "./dice";
 import { GameContext } from "./gameContext";
 import { Player } from "./player";
-import { SpecialSquare, type Square } from "./square";
+import { SpecialSquare, type Square } from "./square/square";
 
 export class Game {
   private boardSize: number;
